+++ example/extended_math_example.dart
@@ -6,4 +6,14 @@ void main() {
 
   // Multiply vectors
   final double res = v1.dot(v2);
+  // Add vectors
+  final Vector res1 = v1.add(v2);
+
+  final v3 = SquareMatrix(<List<double>>[
+    <double>[9, 3, 5], 
+    <double>[-6, -9, 7], 
+    <double>[-1, -8, 1]
+  ]);
+  // Gets determinant of matrix
+  final double det = v3.getDeterminant();
 }