+++ lib/src/exceptions/vector_exception.dart
@@ -1,0 +1,11 @@
+/// An exception that trows when performing improper vector operations.
+class VectorException implements Exception {
+  /// Take [message] that describe reason of exception
+  VectorException(this.message);
+
+  /// Field that contain reason of exception
+  String message;
+
+  @override
+  String toString() => '$message';
+}