+++ lib/src/tensor/tensor.dart
@@ -1,1 +1,7 @@
+import 'base/tensor_base.dart';
 
+/// Class for work with tensors
+class Tensor extends TensorBase<double> {
+  ///
+  Tensor();
+}