+++ test/extended_math_test.dart
@@ -2,7 +2,6 @@ import 'package:extended_math/extended_math.dart';
 
 void main() {
 
-  final v1 = Vector(<double>[1, 2, 3]);
-  final v2 = Vector(<double>[4, 5, 6]);
-  print('${v1.dot(v2)} - 1');
+  final m = SquareMatrix.identity(3);
+  print('${m.data}');
 }