+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# 0.0.1
+# 0.0.2
+
+- Add and implement `add()`, `subtract()` methods to `Vector` class.
+- Add `SquareMatrix` class.
+- Implement `getDeterminant()` methods of `SquareMatrix` class.
+- Add `identity()` constructor for matrices.
+- Move `generate()` constructor from `Matrix` class to `MatrixBase` and add to it
+  `identity` parameter.
+
+## 0.0.1
 
 - Add `Vector`, `Matrix` and `Tensor` classes for working with according objects of linear algebra.
 - Implements addition, subtraction, multiplication methods of `Vector` and `Matrix`.